adequação das saídas dos programas de 1 a 8 e 12 para atender o doctest do curso

+++ Desafio_Devpro/exercicios_EstruturaDeDecisao/exercicio_01_EstruturaDeDecisao.py
@@ -2,8 +2,8 @@
 1 - Faça um Programa que peça dois números e imprima o maior deles.
 
 '''
-numero01 = float(input('Entre com o primeiro número: '))
-numero02 = float(input('Entre com o segundo  número: '))
+numero01 = int(input('Entre com o primeiro número: '))
+numero02 = int(input('Entre com o segundo  número: '))
 
 if numero01 > numero02:
     print(numero01)
